Reworks how credentials are acquired so it uses the aws sdk default chain
The previous assembling of credentials meant that ECS task roles were not able to
be used.  The new method just calls uses the aws sdk's default way of gathering
credentials.  It now just passes the credentials out of that default config
to the sigv4 sign method.  This should provide better flexibility.