Adds functionality to strip out headers from incoming requests
Adds a -s (--strip) flag, which is repeatable. All headers from the incoming
request that match (case insensitive) will be removed.

Adds additional debug logging for this feature.

This is especially useful if you have old sigv2 requests you want to be
able to send through the proxy.  If you do not remove the "Authorization" header
that exists you will get a 400 bad request.