S3-specific non-escaping signer
As S3 URI paths should not be escaped within a signature, the S3
signing case cannot use the same signer instance as the general case
(since URI path escaping is a field of the signer and not a
per-signature parameter).  This adds an `S3Signer` method to the
proxy client and applies it appropriately.

Resolves #11